requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.2.0
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
dataclasses-json==0.6.4
frozenlist==1.4.1
ibm-cos-sdk==2.13.3
ibm-cos-sdk-core==2.13.3
ibm-cos-sdk-s3transfer==2.13.3
ibm-watson-machine-learning==1.0.345
idna==3.6
importlib-metadata==7.0.1
jmespath==1.0.1
jsonpatch==1.33
jsonpointer==2.4
langchain==0.1.5
langchain-community==0.0.17
langchain-core==0.1.18
langsmith==0.0.86
lomond==0.3.3
marshmallow==3.20.2
multidict==6.0.5
mypy-extensions==1.0.0
numpy==1.26.3
packaging==23.2
pandas==1.5.3
pydantic==1.10.13
pydantic_core==2.16.1
python-dateutil==2.8.2
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
rouge==1.0.1
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.25
tabulate==0.9.0
tenacity==8.2.3
tqdm==4.66.1
typing-inspect==0.9.0
typing_extensions==4.9.0
urllib3==1.26.18
yarl==1.9.4
zipp==3.17.0
python-dotenv==0.19.0
ibm_watson==8.3.0
beautifulsoup4==4.10.0
